ssh_api.h: u_int -> size_t

+++ ssh/ssh_api.h
@@ -70,31 +70,31 @@ int	ssh_packet_next(struct ssh *ssh, u_char *typep);
  * the current input packet and the length of this payload.
  * the payload is accessible until ssh_packet_next() is called again.
  */
-void	*ssh_packet_payload(struct ssh *ssh, u_int *len);
+void	*ssh_packet_payload(struct ssh *ssh, size_t *len);
 
 /*
  * ssh_packet_put() creates an encrypted packet with the given type
  * and payload.
  * the encrypted packet is appended to the output byte-stream.
  */
-int	ssh_packet_put(struct ssh *ssh, int type, const char *data, u_int len);
+int	ssh_packet_put(struct ssh *ssh, int type, const char *data, size_t len);
 
 /*
  * ssh_input_space() checks if 'len' bytes can be appended to the
  * input byte-stream.
  */
-int	ssh_input_space(struct ssh *ssh, u_int len);
+int	ssh_input_space(struct ssh *ssh, size_t len);
 
 /*
  * ssh_input_append() appends data to the input byte-stream.
  */
-int	ssh_input_append(struct ssh *ssh, const char *data, u_int len);
+int	ssh_input_append(struct ssh *ssh, const char *data, size_t len);
 
 /*
  * ssh_output_space() checks if 'len' bytes can be appended to the
  * output byte-stream. XXX
  */
-int	ssh_output_space(struct ssh *ssh, u_int len);
+int	ssh_output_space(struct ssh *ssh, size_t len);
 
 /*
  * ssh_output_ptr() retrieves both a pointer and the length of the
@@ -102,12 +102,12 @@ int	ssh_output_space(struct ssh *ssh, u_int len);
  * network. the number of bytes that have been successfully sent can
  * be removed from the output byte-stream with ssh_output_consume().
  */
-void	*ssh_output_ptr(struct ssh *ssh, u_int *len);
+void	*ssh_output_ptr(struct ssh *ssh, size_t *len);
 
 /*
  * ssh_output_consume() removes the given number of bytes from
  * the output byte-stream.
  */
-int	ssh_output_consume(struct ssh *ssh, u_int len);
+int	ssh_output_consume(struct ssh *ssh, size_t len);
 
 #endif
